+++ frmViewMain.pas
@@ -723,6 +723,8 @@ type
     ReachableBuild: Boolean;
     ReferencedBySortColumn: TListColumn;
 
+    FocusedColumnOverride : Integer;
+
     EditInfoCacheLGeneration: Integer;
     EditInfoCache: TArray<string>;
     EditInfoCacheID: Pointer;
@@ -6022,6 +6024,8 @@ var
   i, j, k, l: Integer;
   Rect: TRect;
 begin
+  FocusedColumnOverride := -1;
+
   wbVarPointer := varPointer;
 
   if wbThemesSupported then try
@@ -7108,13 +7112,24 @@ var
   function FoundName: Boolean;
   var
     CellText    : string;
+    i           : Integer;
   begin
     if NameFilter = '' then
       Exit(True);
 
-    CellText := '';
-    vstViewGetText(vstView, aNode, 0, ttNormal, CellText);
-    Result := CellText.ToLowerInvariant.Contains(NameFilter);
+    Result := False;
+    try
+      for i := 1 to Pred(vstView.Header.Columns.Count) do begin
+        CellText := '';
+        FocusedColumnOverride := i;
+        vstViewGetText(vstView, aNode, 0, ttNormal, CellText);
+        Result := CellText.ToLowerInvariant.Contains(NameFilter);
+        if Result then
+          Break;
+      end;
+    finally
+      FocusedColumnOverride := -1;
+    end;
   end;
 
   function FoundValue: Boolean;
@@ -17061,6 +17076,7 @@ var
   ElementCount : Integer;
   i,j          : Integer;
   UseSuffix    : Boolean;
+  FocusedColumn: TColumnIndex;
 begin
   CellText := '';
   NodeDatas := Sender.GetNodeData(Node);
@@ -17073,8 +17089,14 @@ begin
 
   if Column < 1 then begin
 
-    if (vstView.FocusedColumn > 0) and (Pred(vstView.FocusedColumn) <= High(ActiveRecords)) then
-      Element := NodeDatas[Pred(vstView.FocusedColumn)].Element;
+    FocusedColumn := FocusedColumnOverride;
+    if FocusedColumn < 0 then
+      FocusedColumn := vstView.FocusedColumn;
+    if Length(ActiveRecords) = 1 then
+      FocusedColumn := 1;
+
+    if (FocusedColumn > 0) and (Pred(FocusedColumn) <= High(ActiveRecords)) then
+      Element := NodeDatas[Pred(FocusedColumn)].Element;
 
     UseSuffix := Assigned(Element);