+++ whatsnew.md
@@ -2,6 +2,7 @@
 
 ## Bugfixes
 
+* #649 - Right pane filter bug
 * #669 - CELL at 0,0 may be spuriously added to worldspaces during cleaning.
 * #670 - [FO4] Damage magic effects mislabeled subrecord
 * #674 - Invalid ITMs may be reported in rare cases when form versions differ